Move KafkaJsonDecoder code to pinot-json and keep KafkaJsonDecoder class for backward compatability
